refactor: address PR feedback — split large files, remove WorktreeNotFound rejection, restore forkSession
Bridge:
- Split worktree_service.dart into part files (worktree_types.dart, worktree_git_queries.dart)
- WorktreeNotFound no longer causes cleanup rejection (treat missing worktree as 'already cleaned')
- Remove CleanupIssue.worktreeNotFound variant from shared model
- Remove WorktreeNotFound from safety issues and cleanup mapping
- Restore forkSession in OpenCodeApi for future session forking
- Update worktree_service and worktree_cleanup tests

Mobile:
- Split session_list_screen.dart (865 lines) into 4 files under 265 lines each:
  session_list_screen.dart, session_list_actions.dart, session_cleanup_dialogs.dart, session_list_widgets.dart
- Convert all helper function args to named+required
- Remove worktreeNotFound from UI switch and localization

All:
- Add 250-line max file size rule to all 5 AGENTS.md files
