+++ bridge/app/lib/src/bridge/worktree_service.dart
@@ -6,6 +6,9 @@ import "persistence/daos/projects_dao.dart";
 import "persistence/daos/session_dao.dart";
 import "persistence/tables/session_table.dart";
 
+part "worktree_types.dart";
+part "worktree_git_queries.dart";
+
 typedef ProcessRunner =
     Future<ProcessResult> Function(
       String executable,
@@ -15,63 +18,15 @@ typedef ProcessRunner =
 
 typedef GitPathExistsChecker = bool Function({required String gitPath});
 
-// ---------------------------------------------------------------------------
-// Result types
-// ---------------------------------------------------------------------------
-
-sealed class WorktreeSafetyResult {}
-
-class WorktreeSafe extends WorktreeSafetyResult {}
-
-class WorktreeUnsafe extends WorktreeSafetyResult {
-  final List<SafetyIssue> issues;
-  WorktreeUnsafe({required this.issues});
-}
-
-sealed class SafetyIssue {}
-
-class UnstagedChanges extends SafetyIssue {}
-
-class BranchMismatch extends SafetyIssue {
-  final String expected;
-  final String actual;
-  BranchMismatch({required this.expected, required this.actual});
-}
-
-class WorktreeNotFound extends SafetyIssue {}
-
-sealed class WorktreeResult {}
-
-class WorktreeSuccess extends WorktreeResult {
-  final String path;
-  final String branchName;
-  final String baseBranch;
-  final String baseCommit;
-
-  WorktreeSuccess({
-    required this.path,
-    required this.branchName,
-    required this.baseBranch,
-    required this.baseCommit,
-  });
-}
-
-class WorktreeFallback extends WorktreeResult {
-  final String originalPath;
-  final String reason;
-
-  WorktreeFallback({required this.originalPath, required this.reason});
-}
+const _maxWorktreeCreationAttempts = 3;
+const _branchPrefix = "session-";
+const _worktreeDir = ".worktrees";
 
 // ---------------------------------------------------------------------------
 // Service
 // ---------------------------------------------------------------------------
 
 class WorktreeService {
-  static const _maxWorktreeCreationAttempts = 3;
-  static const _branchPrefix = "session-";
-  static const _worktreeDir = ".worktrees";
-
   final ProcessRunner _processRunner;
   final GitPathExistsChecker _gitPathExists;
   final ProjectsDao _projectsDao;
@@ -88,97 +43,16 @@ class WorktreeService {
        _sessionDao = sessionDao;
 
   // -------------------------------------------------------------------------
-  // Git primitives
-  // -------------------------------------------------------------------------
-
-  Future<bool> isGitInitialized({required String projectPath}) async {
-    return _gitPathExists(gitPath: "$projectPath/.git");
-  }
-
-  Future<bool> hasAtLeastOneCommit({required String projectPath}) async {
-    final result = await _runGit(
-      projectPath: projectPath,
-      arguments: const ["rev-parse", "HEAD"],
-    );
-    return result.exitCode == 0;
-  }
-
-  Future<String> resolveDefaultBranch({required String projectPath}) async {
-    final originHeadResult = await _runGit(
-      projectPath: projectPath,
-      arguments: const ["symbolic-ref", "refs/remotes/origin/HEAD"],
-    );
-    final originHeadBranch = _extractBranchName(
-      output: originHeadResult.stdout,
-      prefix: "refs/remotes/origin/",
-    );
-    if (originHeadResult.exitCode == 0 && originHeadBranch != null) {
-      return originHeadBranch;
-    }
-
-    final localHeadResult = await _runGit(
-      projectPath: projectPath,
-      arguments: const ["symbolic-ref", "HEAD"],
-    );
-    final localHeadBranch = _extractBranchName(
-      output: localHeadResult.stdout,
-      prefix: "refs/heads/",
-    );
-    if (localHeadResult.exitCode == 0 && localHeadBranch != null) {
-      return localHeadBranch;
-    }
-
-    final configuredDefaultBranchResult = await _runGit(
-      projectPath: projectPath,
-      arguments: const ["config", "init.defaultBranch"],
-    );
-    final configuredDefaultBranch = configuredDefaultBranchResult.stdout.toString().trim();
-    if (configuredDefaultBranchResult.exitCode == 0 && configuredDefaultBranch.isNotEmpty) {
-      return configuredDefaultBranch;
-    }
-
-    return "main";
-  }
-
-  Future<bool> branchExists({required String projectPath, required String branchName}) async {
-    final result = await _runGit(
-      projectPath: projectPath,
-      arguments: ["branch", "--list", branchName],
-    );
-    return result.stdout.toString().trim().isNotEmpty;
-  }
-
-  Future<ProcessResult> createWorktree({
-    required String projectPath,
-    required String worktreePath,
-    required String branchName,
-    required String baseBranch,
-  }) {
-    return _runGit(
-      projectPath: projectPath,
-      arguments: ["worktree", "add", worktreePath, "-b", branchName, baseBranch],
-    );
-  }
-
-  // -------------------------------------------------------------------------
   // Orchestration
   // -------------------------------------------------------------------------
 
   /// Prepares a worktree for a new or child session.
-  ///
-  /// Returns [WorktreeSuccess] with the path and branch when a worktree is
-  /// ready (either reused from a parent session or freshly created).
-  /// Returns [WorktreeFallback] when the project is not git-initialised, has
-  /// no commits, or every creation attempt fails.
   Future<WorktreeResult> prepareWorktreeForSession({
     required String projectId,
     required String? parentSessionId,
   }) async {
-    // 1. If a parent session exists, reuse its worktree.
     if (parentSessionId != null) {
-      final parentWorktree = await _sessionDao.getSession(
-        sessionId: parentSessionId,
-      );
+      final parentWorktree = await _sessionDao.getSession(sessionId: parentSessionId);
       if (parentWorktree case SessionDto(
         worktreePath: final worktreePath?,
         branchName: final branchName?,
@@ -192,31 +66,19 @@ class WorktreeService {
           baseCommit: parentBaseCommit ?? "",
         );
       }
-      // Parent not found (pre-feature session) — fall through to create new.
     }
 
-    // 2. Guard: must be a git repository.
     if (!await isGitInitialized(projectPath: projectId)) {
       Log.w("WorktreeService: not a git repository: $projectId");
-      return WorktreeFallback(
-        originalPath: projectId,
-        reason: "not a git repository",
-      );
+      return WorktreeFallback(originalPath: projectId, reason: "not a git repository");
     }
 
-    // 3. Guard: must have at least one commit.
     if (!await hasAtLeastOneCommit(projectPath: projectId)) {
       Log.w("WorktreeService: repository has no commits: $projectId");
-      return WorktreeFallback(
-        originalPath: projectId,
-        reason: "repository has no commits",
-      );
+      return WorktreeFallback(originalPath: projectId, reason: "repository has no commits");
     }
 
-    // 4. Resolve base branch and commit before creating the worktree.
-    final baseBranchAndCommit = await resolveBaseBranchAndCommit(
-      projectPath: projectId,
-    );
+    final baseBranchAndCommit = await resolveBaseBranchAndCommit(projectPath: projectId);
     if (baseBranchAndCommit == null) {
       Log.w("WorktreeService: failed to resolve base branch/commit for: $projectId");
       return WorktreeFallback(
@@ -227,18 +89,12 @@ class WorktreeService {
     final baseBranch = baseBranchAndCommit.baseBranch;
     final baseCommit = baseBranchAndCommit.baseCommit;
 
-    // 6. Try to create a worktree, retrying on branch collision.
     for (var attempt = 0; attempt < _maxWorktreeCreationAttempts; attempt++) {
-      final counter = await _projectsDao.incrementAndGetWorktreeCounter(
-        projectId: projectId,
-      );
+      final counter = await _projectsDao.incrementAndGetWorktreeCounter(projectId: projectId);
       final branchName = "$_branchPrefix${counter.toString().padLeft(3, '0')}";
       final worktreePath = "$projectId/$_worktreeDir/$branchName";
 
-      // Skip if branch already exists (counter collision).
-      if (await branchExists(projectPath: projectId, branchName: branchName)) {
-        continue;
-      }
+      if (await branchExists(projectPath: projectId, branchName: branchName)) continue;
 
       final result = await createWorktree(
         projectPath: projectId,
@@ -255,7 +111,6 @@ class WorktreeService {
           baseCommit: baseCommit,
         );
       }
-      // git command failed — try next counter value.
     }
 
     Log.w("WorktreeService: failed to create worktree after 3 attempts for: $projectId");
@@ -265,13 +120,45 @@ class WorktreeService {
     );
   }
 
-  /// Checks whether a worktree is safe to resume (no unstaged changes, correct branch).
+  Future<({String baseBranch, String baseCommit})?> resolveBaseBranchAndCommit({
+    required String projectPath,
+  }) async {
+    try {
+      final storedBranch = await _projectsDao.getBaseBranch(projectId: projectPath);
+      final String baseBranch;
+      if (storedBranch != null && await branchExists(projectPath: projectPath, branchName: storedBranch)) {
+        baseBranch = storedBranch;
+      } else {
+        baseBranch = await resolveDefaultBranch(projectPath: projectPath);
+      }
+
+      final revParseResult = await _runGit(
+        projectPath: projectPath,
+        arguments: ["rev-parse", baseBranch],
+      );
+      if (revParseResult.exitCode != 0) return null;
+
+      final baseCommit = revParseResult.stdout.toString().trim();
+      if (baseCommit.isEmpty) return null;
+
+      return (baseBranch: baseBranch, baseCommit: baseCommit);
+    } on Object {
+      return null;
+    }
+  }
+
+  // -------------------------------------------------------------------------
+  // Safety check
+  // -------------------------------------------------------------------------
+
+  /// Returns [WorktreeSafe] when the directory does not exist — a missing
+  /// worktree is treated as already cleaned up.
   Future<WorktreeSafetyResult> checkWorktreeSafety({
     required String worktreePath,
     required String expectedBranch,
   }) async {
     if (!Directory(worktreePath).existsSync()) {
-      return WorktreeUnsafe(issues: [WorktreeNotFound()]);
+      return WorktreeSafe();
     }
 
     final issues = <SafetyIssue>[];
@@ -295,24 +182,18 @@ class WorktreeService {
       issues.add(BranchMismatch(expected: expectedBranch, actual: actualBranch));
     }
 
-    if (issues.isEmpty) {
-      return WorktreeSafe();
-    }
+    if (issues.isEmpty) return WorktreeSafe();
     return WorktreeUnsafe(issues: issues);
   }
 
-  /// Prunes stale worktree administrative files (best-effort, fire-and-forget).
+  // -------------------------------------------------------------------------
+  // Lifecycle
+  // -------------------------------------------------------------------------
+
   Future<void> pruneWorktrees({required String projectPath}) async {
-    await _runGit(
-      projectPath: projectPath,
-      arguments: const ["worktree", "prune"],
-    );
+    await _runGit(projectPath: projectPath, arguments: const ["worktree", "prune"]);
   }
 
-  /// Removes a worktree. Returns true on success, false on failure.
-  ///
-  /// Calls [pruneWorktrees] first to clean up stale entries, then runs
-  /// `git worktree remove [--force] <worktreePath>`.
   Future<bool> removeWorktree({
     required String projectPath,
     required String worktreePath,
@@ -321,26 +202,12 @@ class WorktreeService {
     if (!_isValidWorktreePath(projectPath: projectPath, worktreePath: worktreePath)) {
       return false;
     }
-
     await pruneWorktrees(projectPath: projectPath);
-
-    final arguments = [
-      "worktree",
-      "remove",
-      if (force) "--force",
-      "--",
-      worktreePath,
-    ];
-    final result = await _runGit(
-      projectPath: projectPath,
-      arguments: arguments,
-    );
+    final arguments = ["worktree", "remove", if (force) "--force", "--", worktreePath];
+    final result = await _runGit(projectPath: projectPath, arguments: arguments);
     return result.exitCode == 0;
   }
 
-  /// Deletes a branch. Returns true on success, false on failure.
-  ///
-  /// Uses `-D` (force) or `-d` (safe) depending on [force].
   Future<bool> deleteBranch({
     required String projectPath,
     required String branchName,
@@ -353,13 +220,6 @@ class WorktreeService {
     return result.exitCode == 0;
   }
 
-  /// Restores (or creates) a worktree at [worktreePath] on [branchName].
-  ///
-  /// If [branchName] already exists, runs `git worktree add <path> <branch>`.
-  /// Otherwise creates a new branch from [baseCommit] (or [baseBranch] if
-  /// [baseCommit] is null) via `git worktree add <path> -b <branch> <ref>`.
-  ///
-  /// Returns true on success, false on failure.
   Future<bool> restoreWorktree({
     required String projectPath,
     required String worktreePath,
@@ -378,73 +238,14 @@ class WorktreeService {
 
     final List<String> addArguments;
     if (verifyResult.exitCode == 0) {
-      // Branch exists — check it out directly.
       addArguments = ["worktree", "add", "--", worktreePath, branchName];
     } else {
-      // Branch does not exist — create from baseCommit (preferred) or baseBranch.
       final startPoint = baseCommit ?? baseBranch;
       addArguments = ["worktree", "add", "-b", branchName, "--", worktreePath, startPoint];
     }
 
-    final addResult = await _runGit(
-      projectPath: projectPath,
-      arguments: addArguments,
-    );
+    final addResult = await _runGit(projectPath: projectPath, arguments: addArguments);
     return addResult.exitCode == 0;
-  }
-
-  Future<({String baseBranch, String baseCommit})?> resolveBaseBranchAndCommit({
-    required String projectPath,
-  }) async {
-    try {
-      final storedBranch = await _projectsDao.getBaseBranch(projectId: projectPath);
-      final String baseBranch;
-      if (storedBranch != null && await branchExists(projectPath: projectPath, branchName: storedBranch)) {
-        baseBranch = storedBranch;
-      } else {
-        baseBranch = await resolveDefaultBranch(projectPath: projectPath);
-      }
-
-      final revParseResult = await _runGit(
-        projectPath: projectPath,
-        arguments: ["rev-parse", baseBranch],
-      );
-      if (revParseResult.exitCode != 0) {
-        return null;
-      }
-
-      final baseCommit = revParseResult.stdout.toString().trim();
-      if (baseCommit.isEmpty) {
-        return null;
-      }
-
-      return (baseBranch: baseBranch, baseCommit: baseCommit);
-    } on Object {
-      return null;
-    }
-  }
-
-  // -------------------------------------------------------------------------
-  // Internals
-  // -------------------------------------------------------------------------
-
-  Future<ProcessResult> _runGit({
-    required String projectPath,
-    required List<String> arguments,
-  }) {
-    return _processRunner("git", arguments, workingDirectory: projectPath);
-  }
-
-  String? _extractBranchName({required Object? output, required String prefix}) {
-    final trimmedOutput = output.toString().trim();
-    if (!trimmedOutput.startsWith(prefix)) {
-      return null;
-    }
-    final branchName = trimmedOutput.substring(prefix.length).trim();
-    if (branchName.isEmpty) {
-      return null;
-    }
-    return branchName;
   }
 
   /// Validates that [worktreePath] is under `<projectPath>/.worktrees/` to
