+++ bridge/sesori_plugin_opencode/lib/src/opencode_api.dart
@@ -202,6 +202,26 @@ class OpenCodeApi {
     return decoded.cast<Map<String, dynamic>>().map(Session.fromJson).toList();
   }
 
+  Future<Session> forkSession({
+    required String sessionId,
+    required String directory,
+  }) async {
+    final client = http.Client();
+    try {
+      final response = await client.post(
+        Uri.parse("$serverURL/session/$sessionId/fork"),
+        headers: {
+          ..._authHeaders,
+          _directoryOpenCodeHeader: directory,
+        },
+      );
+      _ensureSuccess(response, "POST /session/$sessionId/fork");
+      return Session.fromJson(jsonDecode(response.body) as Map<String, dynamic>);
+    } finally {
+      client.close();
+    }
+  }
+
   Future<List<MessageWithParts>> getMessages({
     required String sessionId,
     required String? directory,
