+++ bridge/sesori_plugin_opencode/test/active_session_tracker_test.dart
@@ -803,4 +803,10 @@ class _FakeApi implements OpenCodeApi {
     required String directory,
     required Map<String, dynamic> body,
   }) async => throw UnimplementedError();
+
+  @override
+  Future<Session> forkSession({
+    required String sessionId,
+    required String directory,
+  }) async => throw UnimplementedError();
 }
