+++ bridge/sesori_plugin_opencode/test/opencode_repository_test.dart
@@ -454,4 +454,10 @@ class _FakeApi implements OpenCodeApi {
     required String directory,
     required Map<String, dynamic> body,
   }) async => throw UnimplementedError();
+
+  @override
+  Future<Session> forkSession({
+    required String sessionId,
+    required String directory,
+  }) async => throw UnimplementedError();
 }
