+++ bridge/sesori_plugin_opencode/test/opencode_service_test.dart
@@ -394,6 +394,12 @@ class FakeOpenCodeApi implements OpenCodeApi {
   @override
   Future<ProviderListResponse> listProviders() async =>
       const ProviderListResponse(all: [], defaults: {}, connected: []);
+
+  @override
+  Future<Session> forkSession({
+    required String sessionId,
+    required String directory,
+  }) async => throw UnimplementedError();
 }
 
 class FakeOpenCodeRepository extends OpenCodeRepository {
