+++ mobile/app/lib/features/session_list/session_list_screen.dart
@@ -12,6 +12,10 @@ import "../../core/widgets/app_modal_bottom_sheet.dart";
 import "../../l10n/app_localizations.dart";
 import "rename_session_dialog.dart";
 
+part "session_list_actions.dart";
+part "session_cleanup_dialogs.dart";
+part "session_list_widgets.dart";
+
 class SessionListScreen extends StatelessWidget {
   final String projectId;
   final String? projectName;
@@ -43,14 +47,14 @@ class _SessionListBody extends StatelessWidget {
 
   const _SessionListBody({this.projectName});
 
-  void _goToNewSession(BuildContext context) {
+  void _goToNewSession({required BuildContext context}) {
     final projectId = context.read<SessionListCubit>().projectId;
     context.pushRoute(
       AppRoute.newSession(projectId: projectId),
     );
   }
 
-  void _showSessionActions(BuildContext context, Session session) {
+  void _showSessionActions({required BuildContext context, required Session session}) {
     final loc = context.loc;
     final cubit = context.read<SessionListCubit>();
     final isArchived = session.time?.archived != null;
@@ -79,9 +83,9 @@ class _SessionListBody extends StatelessWidget {
               onTap: () {
                 Navigator.pop(sheetContext);
                 if (isArchived) {
-                  _unarchiveSession(context, cubit, session.id);
+                  _unarchiveSession(context: context, cubit: cubit, sessionId: session.id);
                 } else {
-                  _showArchiveSheet(context, cubit, session);
+                  _showArchiveSheet(context: context, cubit: cubit, session: session);
                 }
               },
             ),
@@ -93,7 +97,7 @@ class _SessionListBody extends StatelessWidget {
               ),
               onTap: () {
                 Navigator.pop(sheetContext);
-                _showDeleteSheet(context, cubit, session);
+                _showDeleteSheet(context: context, cubit: cubit, session: session);
               },
             ),
           ],
@@ -101,267 +105,6 @@ class _SessionListBody extends StatelessWidget {
       },
     );
   }
-
-  // ---------------------------------------------------------------------------
-  // Archive
-  // ---------------------------------------------------------------------------
-
-  void _showArchiveSheet(BuildContext context, SessionListCubit cubit, Session session) {
-    showAppModalBottomSheet<void>(
-      context: context,
-      builder: (_) => _ArchiveSessionSheet(
-        session: session,
-        onConfirm: ({required bool deleteWorktree, required bool deleteBranch}) {
-          _archiveSession(
-            context,
-            cubit,
-            session.id,
-            deleteWorktree: deleteWorktree,
-            deleteBranch: deleteBranch,
-          );
-        },
-      ),
-    );
-  }
-
-  Future<void> _archiveSession(
-    BuildContext context,
-    SessionListCubit cubit,
-    String sessionId, {
-    bool deleteWorktree = true,
-    bool deleteBranch = true,
-    bool force = false,
-  }) async {
-    final loc = context.loc;
-    final success = await cubit.archiveSession(
-      sessionId: sessionId,
-      deleteWorktree: deleteWorktree,
-      deleteBranch: deleteBranch,
-      force: force,
-    );
-    if (!context.mounted) return;
-
-    if (success) {
-      _showUndoSnackBar(context, cubit, loc.sessionListArchived);
-      return;
-    }
-
-    // Check for cleanup rejection (409).
-    final rejection = cubit.lastCleanupRejection;
-    if (rejection != null) {
-      _showForceDialog(
-        context,
-        cubit,
-        sessionId: sessionId,
-        rejection: rejection,
-        isDelete: false,
-        deleteWorktree: deleteWorktree,
-        deleteBranch: deleteBranch,
-      );
-    } else {
-      ScaffoldMessenger.of(context)
-        ..clearSnackBars()
-        ..showSnackBar(SnackBar(content: Text(loc.sessionListArchiveFailed)));
-    }
-  }
-
-  Future<void> _unarchiveSession(
-    BuildContext context,
-    SessionListCubit cubit,
-    String sessionId,
-  ) async {
-    final loc = context.loc;
-    final success = await cubit.unarchiveSession(sessionId);
-    if (!success || !context.mounted) return;
-
-    _showUndoSnackBar(context, cubit, loc.sessionListUnarchived);
-  }
-
-  void _showUndoSnackBar(BuildContext context, SessionListCubit cubit, String message) {
-    final loc = context.loc;
-
-    ScaffoldMessenger.of(context).clearSnackBars();
-    ScaffoldMessenger.of(context)
-        .showSnackBar(
-          SnackBar(
-            content: Text(message),
-            duration: kSnackBarDuration,
-            action: SnackBarAction(
-              label: loc.sessionListUndo,
-              onPressed: () => cubit.undoLastArchiveAction(),
-            ),
-          ),
-        )
-        .closed
-        .then((reason) {
-          // Clear undo state only when the snackbar genuinely expired or was
-          // swiped away. Skip on:
-          //  - action: user pressed undo — cubit already cleared the snapshot.
-          //  - remove: clearSnackBars() was called because a new archive/unarchive
-          //    action replaced this snackbar — clearing now would wipe the *new*
-          //    action's undo state.
-          if (reason == SnackBarClosedReason.timeout || reason == SnackBarClosedReason.swipe) {
-            cubit.clearLastActionUndo();
-          }
-        });
-  }
-
-  // ---------------------------------------------------------------------------
-  // Delete
-  // ---------------------------------------------------------------------------
-
-  void _showDeleteSheet(BuildContext context, SessionListCubit cubit, Session session) {
-    showAppModalBottomSheet<void>(
-      context: context,
-      builder: (_) => _DeleteSessionSheet(
-        session: session,
-        onConfirm: ({required bool deleteWorktree, required bool deleteBranch}) {
-          _deleteSession(
-            context,
-            cubit,
-            session.id,
-            deleteWorktree: deleteWorktree,
-            deleteBranch: deleteBranch,
-          );
-        },
-      ),
-    );
-  }
-
-  Future<void> _deleteSession(
-    BuildContext context,
-    SessionListCubit cubit,
-    String sessionId, {
-    bool deleteWorktree = true,
-    bool deleteBranch = true,
-    bool force = false,
-  }) async {
-    final loc = context.loc;
-    final success = await cubit.deleteSession(
-      sessionId: sessionId,
-      deleteWorktree: deleteWorktree,
-      deleteBranch: deleteBranch,
-      force: force,
-    );
-    if (!context.mounted) return;
-
-    if (success) {
-      ScaffoldMessenger.of(context)
-        ..clearSnackBars()
-        ..showSnackBar(SnackBar(content: Text(loc.sessionListDeleted)));
-      return;
-    }
-
-    // Check for cleanup rejection (409).
-    final rejection = cubit.lastCleanupRejection;
-    if (rejection != null) {
-      _showForceDialog(
-        context,
-        cubit,
-        sessionId: sessionId,
-        rejection: rejection,
-        isDelete: true,
-        deleteWorktree: deleteWorktree,
-        deleteBranch: deleteBranch,
-      );
-    } else {
-      ScaffoldMessenger.of(context)
-        ..clearSnackBars()
-        ..showSnackBar(SnackBar(content: Text(loc.sessionListDeleteFailed)));
-    }
-  }
-
-  // ---------------------------------------------------------------------------
-  // Force delete / archive dialog (409 rejection)
-  // ---------------------------------------------------------------------------
-
-  void _showForceDialog(
-    BuildContext context,
-    SessionListCubit cubit, {
-    required String sessionId,
-    required SessionCleanupRejection rejection,
-    required bool isDelete,
-    required bool deleteWorktree,
-    required bool deleteBranch,
-  }) {
-    final loc = context.loc;
-
-    showDialog<void>(
-      context: context,
-      builder: (dialogContext) {
-        return AlertDialog(
-          title: Text(isDelete ? loc.sessionListForceDeleteTitle : loc.sessionListForceArchiveTitle),
-          content: Column(
-            mainAxisSize: MainAxisSize.min,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(loc.sessionListForceMessage),
-              const SizedBox(height: 12),
-              for (final issue in rejection.issues)
-                Padding(
-                  padding: const EdgeInsets.only(bottom: 4),
-                  child: Row(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Icon(
-                        Icons.warning_amber_rounded,
-                        size: 18,
-                        color: Theme.of(context).colorScheme.error,
-                      ),
-                      const SizedBox(width: 8),
-                      Expanded(child: Text(_describeCleanupIssue(loc, issue))),
-                    ],
-                  ),
-                ),
-            ],
-          ),
-          actions: [
-            TextButton(
-              onPressed: () => Navigator.pop(dialogContext),
-              child: Text(loc.sessionListDeleteConfirmCancel),
-            ),
-            TextButton(
-              onPressed: () {
-                Navigator.pop(dialogContext);
-                if (isDelete) {
-                  _deleteSession(
-                    context,
-                    cubit,
-                    sessionId,
-                    deleteWorktree: deleteWorktree,
-                    deleteBranch: deleteBranch,
-                    force: true,
-                  );
-                } else {
-                  _archiveSession(
-                    context,
-                    cubit,
-                    sessionId,
-                    deleteWorktree: deleteWorktree,
-                    deleteBranch: deleteBranch,
-                    force: true,
-                  );
-                }
-              },
-              child: Text(
-                isDelete ? loc.sessionListForceDeleteAction : loc.sessionListForceArchiveAction,
-                style: TextStyle(color: Theme.of(context).colorScheme.error),
-              ),
-            ),
-          ],
-        );
-      },
-    );
-  }
-
-  String _describeCleanupIssue(AppLocalizations loc, CleanupIssue issue) => switch (issue) {
-    CleanupIssueUnstagedChanges() => loc.sessionListCleanupIssueUnstagedChanges,
-    CleanupIssueBranchMismatch(:final expected, :final actual) => loc.sessionListCleanupIssueBranchMismatch(
-      actual,
-      expected,
-    ),
-    CleanupIssueWorktreeNotFound() => loc.sessionListCleanupIssueWorktreeNotFound,
-  };
 
   @override
   Widget build(BuildContext context) {
@@ -396,7 +139,7 @@ class _SessionListBody extends StatelessWidget {
         ],
       ),
       floatingActionButton: FloatingActionButton(
-        onPressed: () => _goToNewSession(context),
+        onPressed: () => _goToNewSession(context: context),
         tooltip: loc.sessionListNewSession,
         child: const Icon(Icons.add),
       ),
@@ -446,10 +189,10 @@ class _SessionListBody extends StatelessWidget {
                             isArchived: isArchived,
                             isActive: activityInfo != null,
                             backgroundTaskCount: activityInfo?.backgroundTaskCount ?? 0,
-                            onLongPress: () => _showSessionActions(context, session),
+                            onLongPress: () => _showSessionActions(context: context, session: session),
                             onSwipe: () => isArchived
-                                ? _unarchiveSession(context, cubit, session.id)
-                                : _archiveSession(context, cubit, session.id),
+                                ? _unarchiveSession(context: context, cubit: cubit, sessionId: session.id)
+                                : _archiveSession(context: context, cubit: cubit, sessionId: session.id),
                           ); // swipe uses defaults: deleteWorktree=true, deleteBranch=true
                         },
                       ),
@@ -465,401 +208,6 @@ class _SessionListBody extends StatelessWidget {
           onRetry: () => context.read<SessionListCubit>().loadSessions(),
         ),
       },
-    );
-  }
-}
-
-class _SessionTile extends StatelessWidget {
-  final Session session;
-  final bool isArchived;
-  final bool isActive;
-  final int backgroundTaskCount;
-  final VoidCallback onLongPress;
-  final VoidCallback onSwipe;
-
-  const _SessionTile({
-    required this.session,
-    required this.isArchived,
-    required this.isActive,
-    this.backgroundTaskCount = 0,
-    required this.onLongPress,
-    required this.onSwipe,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final loc = context.loc;
-    final updatedAt = session.time?.updated;
-    final filesChanged = session.summary?.files ?? 0;
-
-    return Dismissible(
-      key: ValueKey(session.id),
-      direction: .startToEnd,
-      confirmDismiss: (_) async {
-        onSwipe();
-        // Return false — the cubit removes the item from the list, so we
-        // don't need Dismissible to animate the removal itself.
-        return false;
-      },
-      background: ColoredBox(
-        color: theme.colorScheme.secondaryContainer,
-        child: Align(
-          alignment: .centerLeft,
-          child: Padding(
-            padding: const EdgeInsets.only(left: 24),
-            child: Icon(
-              isArchived ? Icons.unarchive_outlined : Icons.archive_outlined,
-              color: theme.colorScheme.onSecondaryContainer,
-            ),
-          ),
-        ),
-      ),
-      child: ListTile(
-        leading: CircleAvatar(
-          backgroundColor: theme.colorScheme.primaryContainer,
-          child: Icon(
-            Icons.chat_outlined,
-            color: theme.colorScheme.onPrimaryContainer,
-          ),
-        ),
-        title: Text(session.title ?? loc.sessionListUntitled),
-        subtitle: Column(
-          crossAxisAlignment: .start,
-          children: [
-            if (updatedAt != null)
-              Text(
-                loc.sessionListUpdated(_formatTimestamp(updatedAt)),
-                style: theme.textTheme.bodySmall?.copyWith(
-                  color: theme.colorScheme.outline,
-                ),
-              ),
-            if (filesChanged > 0)
-              Text(
-                loc.sessionListFilesChanged(filesChanged),
-                style: theme.textTheme.bodySmall,
-              ),
-            if (isActive)
-              Row(
-                children: [
-                  Icon(Icons.circle, size: 8, color: theme.colorScheme.primary),
-                  const SizedBox(width: 4),
-                  Text(
-                    loc.sessionListRunning,
-                    style: theme.textTheme.bodySmall?.copyWith(
-                      color: theme.colorScheme.primary,
-                      fontWeight: FontWeight.w500,
-                    ),
-                  ),
-                  if (backgroundTaskCount > 0) ...[
-                    Text(
-                      " \u00b7 ",
-                      style: theme.textTheme.bodySmall?.copyWith(
-                        color: theme.colorScheme.primary,
-                      ),
-                    ),
-                    Text(
-                      loc.sessionListBackgroundTasks(backgroundTaskCount),
-                      style: theme.textTheme.bodySmall?.copyWith(
-                        color: theme.colorScheme.primary,
-                        fontWeight: FontWeight.w500,
-                      ),
-                    ),
-                  ],
-                ],
-              ),
-          ],
-        ),
-        isThreeLine: updatedAt != null && (filesChanged > 0 || isActive),
-        trailing: const Icon(Icons.chevron_right),
-        onTap: () {
-          context.pushRoute(
-            AppRoute.sessionDetail(
-              projectId: session.projectID,
-              sessionId: session.id,
-              sessionTitle: session.title ?? "",
-              readOnly: false,
-            ),
-          );
-        },
-        onLongPress: onLongPress,
-      ),
-    );
-  }
-
-  String _formatTimestamp(int ms) {
-    final date = DateTime.fromMillisecondsSinceEpoch(ms);
-    final now = DateTime.now();
-    final diff = now.difference(date);
-
-    if (diff.inMinutes < 1) return "just now";
-    if (diff.inHours < 1) return "${diff.inMinutes}m ago";
-    if (diff.inDays < 1) return "${diff.inHours}h ago";
-    if (diff.inDays < 30) return "${diff.inDays}d ago";
-    return "${date.year}-${date.month.toString().padLeft(2, "0")}-${date.day.toString().padLeft(2, "0")}";
-  }
-}
-
-class _StaleProjectView extends StatelessWidget {
-  final VoidCallback onBack;
-
-  const _StaleProjectView({required this.onBack});
-
-  @override
-  Widget build(BuildContext context) {
-    final loc = context.loc;
-
-    return Center(
-      child: Padding(
-        padding: const EdgeInsets.all(24),
-        child: Column(
-          mainAxisSize: .min,
-          children: [
-            Icon(
-              Icons.folder_off_outlined,
-              size: 48,
-              color: Theme.of(context).colorScheme.error,
-            ),
-            const SizedBox(height: 16),
-            Text(
-              loc.sessionListStaleProjectTitle,
-              style: Theme.of(context).textTheme.titleMedium,
-            ),
-            const SizedBox(height: 8),
-            Text(loc.sessionListStaleProjectMessage, textAlign: .center),
-            const SizedBox(height: 24),
-            FilledButton.icon(
-              onPressed: onBack,
-              icon: const Icon(Icons.arrow_back),
-              label: Text(loc.sessionListStaleProjectBack),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class _ErrorView extends StatelessWidget {
-  final ApiError error;
-  final VoidCallback onRetry;
-
-  const _ErrorView({required this.error, required this.onRetry});
-
-  @override
-  Widget build(BuildContext context) {
-    final loc = context.loc;
-
-    return Center(
-      child: Padding(
-        padding: const EdgeInsets.all(24),
-        child: Column(
-          mainAxisSize: .min,
-          children: [
-            Icon(
-              Icons.error_outline,
-              size: 48,
-              color: Theme.of(context).colorScheme.error,
-            ),
-            const SizedBox(height: 16),
-            Text(
-              loc.sessionListErrorTitle,
-              style: Theme.of(context).textTheme.titleMedium,
-            ),
-            const SizedBox(height: 8),
-            Text(_describeError(loc, error), textAlign: .center),
-            const SizedBox(height: 24),
-            FilledButton.icon(
-              onPressed: onRetry,
-              icon: const Icon(Icons.refresh),
-              label: Text(loc.sessionListRetry),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  String _describeError(AppLocalizations loc, ApiError error) => switch (error) {
-    NotAuthenticatedError() => loc.apiErrorNotAuthenticated,
-    NonSuccessCodeError(:final errorCode, :final rawErrorString) =>
-      rawErrorString != null
-          ? loc.connectErrorNonSuccessCodeWithBody(
-              errorCode,
-              rawErrorString,
-            )
-          : loc.connectErrorNonSuccessCode(errorCode),
-    DartHttpClientError(:final innerError) => loc.connectErrorConnectionFailed(innerError.toString()),
-    JsonParsingError() => loc.connectErrorUnexpectedFormat,
-    GenericError() => loc.connectErrorUnknown,
-  };
-}
-
-// -----------------------------------------------------------------------------
-// Delete session bottom sheet
-// -----------------------------------------------------------------------------
-
-class _DeleteSessionSheet extends StatefulWidget {
-  final Session session;
-  final void Function({required bool deleteWorktree, required bool deleteBranch}) onConfirm;
-
-  const _DeleteSessionSheet({required this.session, required this.onConfirm});
-
-  @override
-  State<_DeleteSessionSheet> createState() => _DeleteSessionSheetState();
-}
-
-class _DeleteSessionSheetState extends State<_DeleteSessionSheet> {
-  bool _deleteWorktree = true;
-  bool _deleteBranch = true;
-
-  @override
-  Widget build(BuildContext context) {
-    final loc = context.loc;
-    final theme = Theme.of(context);
-
-    return Padding(
-      padding: const EdgeInsets.fromLTRB(16, 8, 16, 16),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: [
-          Text(
-            loc.sessionListDeleteConfirmTitle,
-            style: theme.textTheme.titleMedium,
-          ),
-          const SizedBox(height: 8),
-          Text(
-            loc.sessionListDeleteConfirmMessage,
-            style: theme.textTheme.bodyMedium?.copyWith(
-              color: theme.colorScheme.onSurfaceVariant,
-            ),
-          ),
-          const SizedBox(height: 12),
-          CheckboxListTile(
-            value: _deleteWorktree,
-            onChanged: (v) => setState(() => _deleteWorktree = v ?? false),
-            title: Text(loc.sessionListDeleteWorktreeCheckbox),
-            dense: true,
-            controlAffinity: ListTileControlAffinity.leading,
-            contentPadding: EdgeInsets.zero,
-          ),
-          CheckboxListTile(
-            value: _deleteBranch,
-            onChanged: (v) => setState(() => _deleteBranch = v ?? false),
-            title: Text(loc.sessionListDeleteBranchCheckbox),
-            dense: true,
-            controlAffinity: ListTileControlAffinity.leading,
-            contentPadding: EdgeInsets.zero,
-          ),
-          const SizedBox(height: 12),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.end,
-            children: [
-              TextButton(
-                onPressed: () => Navigator.pop(context),
-                child: Text(loc.sessionListDeleteConfirmCancel),
-              ),
-              const SizedBox(width: 8),
-              FilledButton(
-                style: FilledButton.styleFrom(backgroundColor: theme.colorScheme.error),
-                onPressed: () {
-                  Navigator.pop(context);
-                  widget.onConfirm(
-                    deleteWorktree: _deleteWorktree,
-                    deleteBranch: _deleteBranch,
-                  );
-                },
-                child: Text(loc.sessionListDeleteConfirmAction),
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-// -----------------------------------------------------------------------------
-// Archive session bottom sheet
-// -----------------------------------------------------------------------------
-
-class _ArchiveSessionSheet extends StatefulWidget {
-  final Session session;
-  final void Function({required bool deleteWorktree, required bool deleteBranch}) onConfirm;
-
-  const _ArchiveSessionSheet({required this.session, required this.onConfirm});
-
-  @override
-  State<_ArchiveSessionSheet> createState() => _ArchiveSessionSheetState();
-}
-
-class _ArchiveSessionSheetState extends State<_ArchiveSessionSheet> {
-  bool _deleteWorktree = true;
-  bool _deleteBranch = true;
-
-  @override
-  Widget build(BuildContext context) {
-    final loc = context.loc;
-    final theme = Theme.of(context);
-
-    return Padding(
-      padding: const EdgeInsets.fromLTRB(16, 8, 16, 16),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: [
-          Text(
-            loc.sessionListArchiveConfirmTitle,
-            style: theme.textTheme.titleMedium,
-          ),
-          const SizedBox(height: 8),
-          Text(
-            loc.sessionListArchiveConfirmMessage,
-            style: theme.textTheme.bodyMedium?.copyWith(
-              color: theme.colorScheme.onSurfaceVariant,
-            ),
-          ),
-          const SizedBox(height: 12),
-          CheckboxListTile(
-            value: _deleteWorktree,
-            onChanged: (v) => setState(() => _deleteWorktree = v ?? false),
-            title: Text(loc.sessionListDeleteWorktreeCheckbox),
-            dense: true,
-            controlAffinity: ListTileControlAffinity.leading,
-            contentPadding: EdgeInsets.zero,
-          ),
-          CheckboxListTile(
-            value: _deleteBranch,
-            onChanged: (v) => setState(() => _deleteBranch = v ?? false),
-            title: Text(loc.sessionListDeleteBranchCheckbox),
-            dense: true,
-            controlAffinity: ListTileControlAffinity.leading,
-            contentPadding: EdgeInsets.zero,
-          ),
-          const SizedBox(height: 12),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.end,
-            children: [
-              TextButton(
-                onPressed: () => Navigator.pop(context),
-                child: Text(loc.sessionListDeleteConfirmCancel),
-              ),
-              const SizedBox(width: 8),
-              FilledButton(
-                onPressed: () {
-                  Navigator.pop(context);
-                  widget.onConfirm(
-                    deleteWorktree: _deleteWorktree,
-                    deleteBranch: _deleteBranch,
-                  );
-                },
-                child: Text(loc.sessionListArchiveConfirmAction),
-              ),
-            ],
-          ),
-        ],
-      ),
     );
   }
 }
